src/requirements.txt
flask==2.0.1
psycopg2==2.9.1
pyjwt==2.1.0
bcrypt==3.2.0
gunicorn==20.1.0
